"Done with the paper. As expected I do not understand it, but it is a very well and clearly written paper. It has convinced me – just like last year I tried that thing where I trained a convolutional net on the 15-puzzle problem, I need to dive deeper into partial evaluation.
I'll look a little into embedding and after that my next goal for the language will be to derive KMP via specialization in my own language. If I can master the stuff in this paper, I will have made another step forward. I was thinking about union types, but I had no idea that pattern matching and string matching could be related. The reason that they are is why going for string matching is so attractive to me right now.

Going for pattern matching off the bat seems complicated, but with some focused effort I just might be able to do string matching.

I needed to put in PE capabilities for primitive operations into the language at some point and now seems a good time as any to do it.
"

The paper I am talking about in the above excerpt is this one:

http://www.brics.dk/RS/03/20/BRICS-RS-03-20.pdf
Fast Partial Evaluation of Pattern Matching in Strings

It is really great how now I have access to scientific articles on the Internet. It really makes a world of difference to me. I could not have done this kind of thing back in high school.

At any rate, now I know what my next goal with the regards to the language will be.

I won't be able to run the code directly unless I extract it, but I will be able to tell the quality just by looking at it.

https://www.youtube.com/watch?v=iZ93Unvxwtw

Here is a lecture by Sedgewick that explains the KMP algorithm.