I am done studying those PE papers on string matching. I've also looked at embedding. I'll just do it like Python and Lua except better since I will be calling Spiral from F#. Forget about that that. It is not worth anymore brain cycles.
The real difficult part will soon come.

Unifying pattern matching, partial evaluation and union types will be my greatest achievement and it will be done by plugging away at it just as usual.

Just like how I finally resolved the question of how to properly return functions by properly designing their node and type, so will there turn out to be that there is an exquisite simplicity hiding here.

I am willing to spend the rest of the year on this if necessary until I get it right.

In hindsight I suppose I should have studied partial evaluation instead of type systems in February, but even a negative result is still information.

The tasks ahead of me would be to go through the 1987 book on implementing functional languages by SPJ and then I will start work on deriving the efficient string matcher from a naive one using partial evaluation.

Also I need to give the book by Jones a second chance. I will study it more in depth once it is time to get back to work on coding.

Apart from that, everything from the field of partial evaluation is worth studying and I will make it a resolution to follow the field from here on out. I've already started that in a way.

I will read a bit of the SPJ book and then call it a day.