As expected it will take me a while before I return to programming.
Right now I am on the page 82/458 of the SPJ book.

I've made one decision and that is to reset the development on the language. I'll rewrite it from scratch, this time with type constructors and direct pattern matching and I'll replace the Cuda backend with an F# one.

Putting in sum types will not be a light undertaking, but I need to do it, and I won't be able to do proper testing with Cuda code.

After I am done, I will merge all the pieces together into one language and Spiral will be complete.

The reason why I need to do it from the start is the same as the previous 10 times I did this, without the resolve to do it like this, to demolish what I built previously, I will be emotionally bound to the code I did previously and I won't be able to do proper work.

This is how things go. You build it up and then tear it down and every time it is built up, it comes out much better than before.

I need to forget wha I did up to now, so I can focus on the future.

Of course, I will reuse code from the old versions as always which will speed things up significantly, but I will not be bound by it.

With that said, let me get back to the book. I've been going through it at a grinding pace, but these chapters on pattern matching are exactly what I need.