I rewrote the example from the SPJ book and studied it for a while. I would not say that I've internalized pattern matching completely, but I do have an idea of what it the compiler is trying to do now. So I've internalized a part of it.
I need to study it for a bit more, but I think I understand the PM compiler well enough to implement it in my own language now.

...

Hmmmmm...should I? What should be my plan from here? This is what I am thinking of right now.

How do I fit this with the rest of my plans?

The thing I am leaning towards is to put this in. Copy the old version and then plug this in and then make the F# backend so I can play with this.

No matter what, I do not think I should really start off with patial evaluation. I have no idea how to do it that way anyway, but I get the feeling that building in the pattern matching compiler would be the first correct step towards this.

Once I compile the match to a case, I should be able to optimize the expressions using partial evaluation so how about I am for that?