"3:40pm. and Pattern =     | A of Pattern * string // Type annotation case     | A' of Pattern * Ty     | S of string     | S' of string // match if not tuple     | R of Pattern list * Pattern option // Tuple     | F of Pattern * string // Functiona application with retracing.     | N of string * Pattern // matches a tuple name and proceeds onto the pattern on a hit. I made up my mind to ignore everything that I have done up to now, so I've erased everything past this.
...
type Ty =
    | PrimT of SpiralDeviceVarType
    | VVT of Ty list * string
    | FunctionT of EnvTy * FunctionCore
    | ModuleT of EnvTy
    | LocalPointerT of Ty
    | SharedPointerT of Ty
    | GlobalPointerT of Ty
    | ClosureT of Ty * Ty
    | ForCastT of Ty // For casting type level function to term (ClosureT) level ones.
    | ForModuleT of string

3:45pm. I am thinking of how I should do the type constructor types. Should it be a name? A variable...no not a variable. I just managed to purge them from the type level at last and there is no need to repeat the mistakes of the past three months. I know better than that now.

My intuition is telling me that unlike in ML style languages, I should be able to make type constructors functions if I play my cards right.

3:55pm. I am still thinking. Let me get some juice. I keep forgetting that there is pear juice in the fridge.

...I forgot the juice. I got distracted by the bug repelant.

4:05pm. I am still thinking.

About the semantics of named tuples...

| ForModuleT of string

Is this thing really necessary? I know why I did it, but there might be a better way.

I know that I made the .Name(fields)  notation for named tuples and .name for module application, but an idea just occured to me.
    | N of string * Pattern // matches a tuple name and proceeds onto the pattern on a hit.

Consider this thing here.

Why could I make name extraction and name tuple creation separate.

If I write something like `f .SetF (2,3)`, instead of that being two things, how about I make them three. SetF can stand on its own and work as it does in pattern matching.

The way to make named tuples with be to apply 2,3 to .SetF as if it was a function. That is the way to do this. That way I can replace ForModuleT with something more sensible and decrease the distance between functions and modules. They should really be the same thing.
| NameT of string

4:20pm. I am still focusing. Yeah, I am recaling all of the thinking I did on how to do union types.

4:25pm. This is good. I haven't written much code today, apart from translating that thing in the book which comes to around 100 lines so it is not too bad...but for the past hour I've been super focused.

type W =
| A
| B of int
| C of W * W

So how should I represent something like the above? Inwardly I want it to be an union type wrapped in  type of W. So something like

`{A, B of int, C of rect_0 * rect_0}` where that whole expression is rect_0. And the stuff between {} is an ordered set.

Now A, B and C...I think I should represent them nominally much like named tuples so in fact they should be just strings.

In addition, when each type is defined it should export functions named A, B and C as constructors for the underlying types.

And unlike in F# and like in ML and Haskell, those constructors should be first class. Rather than use tuples for everything, I should really be using functions as they are much more convenient. This is something F# got wrong and probably for compatibility reasons. I do not care about that.

Now, if I write something like `B 2` what should the object returned be?

Well, it should be similar to a tuple, but exactly a tuple...nah. A function? Not that it either. Neither can it be an union since those are just sets.

It should be something like (B 2, rect_0).

| RecT of int

So let me add this for starters to the types.

Raw expression or int...lunch.

4:55pm. Done with lunch. That defocused me a bit. What I want to say is whether RecT should be a raw expression like a function or an int index. I am leaning towards index. There are definite benefits towards keeping functions the way they are, but using that for types...

Nah, let me just use ints.

5:05pm. Before I can do anything there is just one final thing left...

That is to formalize how the discriminated unions should be represented.

...Really, couldn't I just reuse tuples for this?

Before I do that, let me just add one thing.
| UnionT of Set<Ty>

5:10pm.
| TypeConstructor of Ty list * string

And this thing too. Type constructors really are quite similar to tuples. I definitely see the similarity now.

Now how does TyVV look like again? I might be able to reuse it since the mechanisms are so similar...
| TyVV of TypedExpr list * Ty

Yeah, I really should just reuse TyVV in the typechecker and not worry about it too much. This is it. This is how it should be done.

5:20pm. The only difference from an actual tuple is that the type will be sealed.

...With that most of the pieces are in place.

The only things left to decide is how types are going to be memoized. I am honestly leaning towards just using the standard memoDict for this...though then I won't be able to index by int.

Well, the standard MemoDict is the proper choice for actually memoizing the results of type application. It is what I need, so I will have to add a TypeDone case. Unlike with methods, I will not have to do any type guards, on recursion I will just assign a new integer to the type. As planned, types will be constructed using functions. This is good as it will allow me to reuse the general mechanism.

For actually unfolding the recursive types which will be necessary in pattern matching, I will add the TypeDict<int,Ty>. I do need a second dictionary just for this and it will serve me nicely.

5:35pm. Hmmm...yeah, it all checks out in my head now. I feel like I have a model of next version of the typechecker. Two hours of thinking, straight from my brain. It feels like I did good work. I really got into the zone. I do not feel like coding right now, I'll do it tomorrow.

Of course, there are still some bits missing, I haven't a 100% of what should be done in my head, but the missing pieces are not difficult and I'll fill them as I work along. The real hard part was the design work here.

Implementing all of this in the partial evaluator (typechecker) and the F# backend will take me a while, but I feel the road is clear and can cross it now. If I start walking nothing will run me over. This was not just done in the last two hours of course, I've been assembling the pieces for months and it is finally time to put them together.

Tomorrow I can start coding.

Discriminated unions will be Spiral's last great feature – with it I will have what it takes to clear away the obstacles in front of me and complete the library. Maybe I will kill C++ while I am at it."

The above excerpt from my journal says it all. 5:45pm is hardly the time to start a new coding session.