"10:20am. But really I feel dry.
Just for a bit here, I was in the zone. I think I understand how to deal with union types.

I need 3 different operations in my language to make pattern matching work.

1) Destructuring I already have and know how to make work so I won't go into it.
2) Splitting – This is what I would do on an union type. I split on it and supposing the union has 3 separate types T1,T2,T3, then that turns into 3 cases T1 -> E1, T2 -> E2, T3 -> E3.
3) Unfolding – That is when I take a recursive type and unfold it a level.

I am not sure what the semantics should exactly be for #3. For non-recursive union types it would be much easier since I only would need to add splitting.

10:25am. Oh, yes, it might not be suspicious with the way I am going about this, but I won't use the algorithm from the book straightforwardly.

The problem is just that the way ML styled language and Spiral approach pattern matching is just too different. It might be worth looking into how pattern matching is implemented in Racket and Scheme instead.

The way it is done in the SPJ book is that the pattern matcher is compiled directly. The way I would like to do it is to iteratively split and deepen until exhaustiveness or lack of thereof.

10:50am. I just do not see it. Maybe that split and unfold algorithm is related to the automata thing I've seen mentioned in the pattern matching papers.

But the thing is, how exactly do I refold it so to speak...

By that I mean...

Suppose I have something like...
let x = function
    | a :: b :: c :: d -> 1
    | a -> 2
If I split on this, I will get a large number of useless cases.

I will get...

a :: b :: c :: d -> 1
a :: b :: c -> 2
a :: b -> 2
a -> 2

Doing it this way would be easy to guarantee exhaustiveness and get rid of unused pattern matching cases which is why I want to use it instead of the one by Wadler. I also want a single match to be applicable to multiple patterns.

As interesting as the algorithm is, I do not see a way to adapt it to my needs, but I see what its point is.

The above example is actually much worse since each variable will also be split into Cons and [], meaning I will have to check 16 cases in total instead of 2 unless I figure out how to optimize it. The way I would do it now, the number of cases would be n^d where n is the number of types in a union and d is the unfolding depth.

Yikes.

11:05am. What I need during a pattern match success is to renormalize the type. That would turn the exponential case to a linear one. In essence it would turn the algorithm for a brute force depth first search to a dynamic programming one. Memoization is always the answer to everything.

11:25am. ...I am deep in thought.

I think I know how to do it. I do not need to renormalize the types. I just have to make the pattern matcher grow the case tree in the right way."

I need a different algorithm than in the book as the above excerpt indicates. I am not at all motivated to start coding yet.