``` and TypedPattern =     | PatS of TyV     | PatR of TypedPattern list * TypedPattern option     | PatF of TypedPattern * TypedExpr     | PatWhen of TypedPattern * TypedExpr ```
A bit by bit, I am separating the task into relevant pieces. Written out as above, it does not seem like it would require much insight, but it did take me quite a bit of thinking to come to.

Somewhere along the way it did slip my mind that I do not have to do everything in the partial evaluator. I can just check for which patterns are valid and assign proper variables to them.

Once I do that, I've already done half the pattern matching job and if I want to I can translate them to F# directly or naively to C.

As a programmer, I always go through the same pattern where I try to carry too large of a burden and it stresses me. The end result of that is always simplification. I keep looking for ways to lighten up and I find them.

At the end of suffering lies inspiration.

https://ocw.mit.edu/courses/electrical-engineering-and-computer-science/6-001-structure-and-interpretation-of-computer-programs-spring-2005/video-lectures/4a-pattern-matching-and-rule-based-substitution/

Now that I've gotten that innovation down, I still feel like reading a bit more on pattern matching compilation. I was not aware that there was stuff on this in the SICP book. I must have missed it while skimming the thing. The above is video from the course.

Instead trying to wrack my brain on how I am going to deal with Pattern and such, reasoning about TypedPattern in comparison will be 10x easier and will allow me to make headway in my understanding of pattern matching algorithms.

Once I have a matcher in such a form, the only thing left that remains is to merge the patterns together in addition to reordering them.