"11am. I had to lie down for a bit to think. I think I mostly worked out the pattern matcher.
Let me just check the journal so I can recollect my thoughts.

11:05am. Ah, here it is, let me start from here. I wrote this a few days ago.

"I need 3 different operations in my language to make pattern matching work.

1) Destructuring I already have and know how to make work so I won't go into it.
2) Splitting – This is what I would do on an union type. I split on it and supposing the union has 3 separate types T1,T2,T3, then that turns into 3 cases T1 -> E1, T2 -> E2, T3 -> E3.
3) Unfolding – That is when I take a recursive type and unfold it a level."

I wrote a lot of stuff in the last few days, and even when I got close to the answer it was not enough to quite enough to grip the problem by the handle. It was a constant struggle to keep the model from slipping.

What I want to say is that the above 3 pieces have a point, but it is not the point I need. It is not the correct, optimal cut.

What I have to do in order to make the pattern matching model sensible is keep 3 things in mind:

1) Static -> static transitions. These kinds of things can be optimized. The pattern matching that is currently in the language only does this.
2) Static -> dynamic transitions. These ones are where the typed pattern matching construct should be instantiated. This is the boundary.
3) Dynamic -> dynamic transitions. These transitions should produce nested terms.

There is also the fourth transition, but I'll save that for the ascent to transcendence.

Basically, all of the problems that I have right now are stemming from the drive to push the above 3 things into just two or even one box. It is somewhat understandable that I might be confused. If I think of how regular arithmetic operations should be partially evaled then static -> static can be thought as one thing and static -> dynamic can be folded into dynamic -> dynamic.

It makes sense, the scenarios for everything but pattern matching are binary – either they can be pevaled or they can't.

Pattern matching on the other hand has a significant boundary and I've failed thus far to properly draw it out."

I am almost there. It feels like I am going to deal iwth both pattern matching and partial evaluation in one feel swoop.

Days ago, I thought I had a problem with figuring out pattern matching algorithms, but that is not it at all.

What I really needed was not to optimize those things, but to draw up the boundaries instead. That is where the difficulty lies.

I do not feel like starting just yet. Let me study PE for a bit more.