"I need to figure out just this single thing before I can start.
It is all fine to optimize single expressions, but what if I need to match something like...

| C(2,y) -> E1
| C(x,3) -> E2

with the static dtat C(q,w)

According to my scheme, since we know the constructor is C, I can take that out.

Now I can just match tuple 2,y and x,3, but here is the thing.

The first match could be compiled like...
match q with
| 2 -> E1

and the second would be:

match w with
| 3 -> E2

Now the way I intended this is to merge the match clauses.

I'd really want to do something like

match q, w with
| ...

instead, but it is obvious that in the above example that would not be possible.

2:15pm. So should I just nest them like

match q with
| 2 -> E1
| _ -> match w with
 | 3 -> E2

And unlike what I thought, I can see that a single pattern for a match with a dynamic component would not be enough.

2:20pm. Yeah, let me just nest them and I'll optimize that away if the both match variables are the same. That is the smart way to do it. If I have something like...

match q with
| 1 -> E1
| _ -> match q with
 | 2 -> E2

I'll just have the compiler turn that into...
match q with
| 1 -> E1
| 2 -> E2

Makes sense right. Really, the biggest problem in my language for pattern matching is figuring out how to overcome the dynamic static boundary.

I'll need to nest in order to implement active patterns anyway.

2:25pm. Like in the ML paper, I will need to implement Yes, No, and Maybe for the result of pattern matching. If a split can only match a partial result then I will need to return maybe.

...Ahhh, what a bother.

All of this, all of this should be optimized with partial eval directly, but I am not sure how to do it!

Agh.

2:30pm. Since it has come to this, how about I just decide to nest everything?

Yeah, doing it naively might make it easier to apply partial evaluation later once I figure out how that is supposed to be done.

But if I am going to apply partial evaluation wouldn't I need something to destructure cases directly.

Yeah, how about I get rid of nested patterns and instead compile the case expressions by nesting them? I could use the same thing as in the SPJ book.

Ok.

2:35pm.
and TypedPattern =
    | PatV of Value
    | PatS of TyV
    | PatR of string * TypedPattern list * TypedPattern option
    | PatC of string * TypedPattern list * TypedPattern option
    | PatF of TypedPattern * TypedExpr
    | PatWhen of TypedPattern * TypedExpr

Let me change the above to...
/// Typed patterns cannot be statically evaluated.
and TypedCase =
    | PatR of string * TyV list
    | PatC of string * TyV list

...But maybe I am wrong. Maybe I really should be compiling the Pattern to a case.

...And I should be doing that in the prepass, before the function is even ran.

In the first place, I definitely cannot apply peval to Pattern. Cearly they must be Expr, so the interpreter itself has the chance to consider the cases, isn't that right?

Really, it would be the easiest if every tuple, every type constructor and every type had its own tag wouldn't it?

That I could do pattern matching at runtime.

The truth is, I wanted to optimize the tags for every pattern and make them sequential, but for the first try, why don't I make everything dynamic to start things off like in Lisp?

Well, no, the primitive types definitely do not need tags since I will have the exact constructors described, but the rest is fair game right?

No, actually, I will need tags for primitive types in union types. The primitive types in tuples and constructors won't need it though.

2:50pm. Actually, I can wrap the primitive types in a singleton nameless tuple and be done with that.
and TypedCase =
    | PatR of string * TyV list * TyV option
    | PatC of string * TyV list

How about I do it like this. This makes the most sense right. Active pattern and the rest would work best in terms of standard langauge constructs.

...No, wait, wait, wait.

It is the untyped case that needs to be like this.
and Case =
    | CaseR of string * string list * string option
    | CaseC of string * string list

/// Typed patterns cannot be statically evaluated.
and TypedCase =
    | TCaseR of string * TyV list * TyV option
    | TCaseC of string * TyV list

So something like this?

2:55pm. Also when the If branch can be statically determined, I think I will waive the requirement for the types of both branches to be equal in order to ease up on compiling the When case.

3pm. Yeah, this is the path of radical simplification. Do not even have the functions as patterns, just compile them to Expr and then do partial evaluation to select the correct case.
| TyCase of TypedCase * TypedExpr * TypedExpr * Ty

Yeah, this is the way to do this.
| Case of Case * Expr * Expr

The Case will be analogous to the if-then-else.

I won't bother with the efficiency of it.

Having to go over clauses makes everything 100x more complicated.

And this sets the stage for partial evaluation nicely."

As usual the stuff in quotes is from my journal.

I am just going round and round the language design landscape right now. Right now I am leaning towards getting rid of Pattern from function signatures. I'll make a compiler to turn Pattern into Expr.

Let me think about it some more.