I spent like 3 weeks translating Petzold's GUI examples to F#.
Maybe I should make a resolution and translate the partial evaluators from the book into F#.

The main reason why the hell it took me so long to decide to turn Patterns into Exprs is because of the lack of knowledge of how to optimize them.

That is still the case in fact. Yeah, so let me do that.

It is not that there is such a huge number of necessary things in computer science. Partial evaluation will be worth it.

I'll resume work on this repo after I finish doing that. I'll open the `PE stuff` repo when it is time to start.