"1:15pm. 97/425.
...This is really quite different than I expected. I totally do not get this. I wish they'd given the mix algorithm in ML rather than pseudocode.

I get that is is doing the memoization, but I do not at all feel they are using the same approach as me.

1:25pm. An idea comes to me. There is something I can do in order to enable term rewriting (which is related to common sub-expression elimination.)

Suppose I have something like this:

inl x = get_tuple q 1
inl y = get_tuple q 1

The way this would work now is that x and y would get assigned the same types, but different tags.

Instead if I had something like:

inl x = memo_expr (get_tuple q 1)
inl y = memo_expr (get_tuple q 1)

Then x and y would get assigned the same tags!

Yeah, this is it! Without a doubt, this is the way to propagate information correctly. In a purely functional language, I could apply this optimization to every expression. Here I'd either have to do it manually or build in a separation of effects. For just optimizing the naive pattern matcher, the manual way would work nicely.

What memo_expr would do it is evaluate the expression and then bind that to the destructuring of that expression. Then when memo_expr is called again with the same expression it would look in the cache return the original destructured variable.

Memoization really is the answer to everything.

All the `if!!` statetement woudl have to do is replace the conditional with true or false in the branches and the information would propagate perfectly. I'd have to introduce `memo_expr_assume` for internal use, but because the `memo_expr` operation is composable I can get what I want by nesting it wherever I want."

As usual the stuff in quotes is from my journal.

It is no thanks to the book, but I finally figured out how optimize pattern matchers directly in the language.

Now that I've written it out like this, it seems so trivial, like even a beginner could have come up with it, but this bit of insight took me like weeks to arrive at.

I'll get back to reading. I opened the `Partial Evaluation Book Stuff` repo and translated the first three examples.

I'll finish the book in a few days, but I am not sure for how long I will study it after that. I wish the mix algorithm was given in ML instead of pseudocode.