So I've decided to change the evaluation strategy from call by value to call by value + local scope expression memoization.
Inspired by the book, I'll also inline function to which all arguments are static.

I am currently 1/3rd into the book, but before I do anything even after I finish it, I think I will dedicate some serious time to studying term rewriting and the things related to it like common subexpression elimination.

The new evaluation strategy has some similarities to term rewriting so I want to research this a bit before commiting to anything.

But at this point, I think that all the steps to making the pattern matching compiler and optimizing it natively are set.

I wrote the update 3 days ago, but I had not even suspected that the way to solve all of this is to change the evaluation strategy so it does more memoization.

Programming is exactly like channeling will.

I put in effort and new pathways get created in my brain that unlock new features in the programs I am making.

Today, I did not do much as I partied yesterday and was recovering from that, but tomorrow hopefully I will make some headway into the book.

I only figured out how to the pattern matcher while I was in the middle of it, and now that I have this understanding, I no longer really need it so I'll finish it in 2-3 days.

Tomorrow I will also proof read the update, extend it a bit and post it. It is time.