Done with the PE book. Moving to studying term rewriting.
Fuu, 425 pages... To be honest, as most of my studies this stuff was 90% useless, and I generally wasted my time.

I think I will take a break now and then start reading up on term rewriting. I have that big scary book already sitting on my hard drive from a while ago and now that I see the potential of it a bit better, I want to make use of it.

I am going to ask around for what call by value + memoization is supposed to be on the PL after I ease myself in.