I've finally started work on the new partial evaluator/typechecker. The next on the list is the destructure function, but let me leave it alone for today.
For the past two weeks, the only things I did were think and do research. In terms of code produced, I might not have accomplished much, but the effort was very much real.

The wheat has grown and now it is time for the harvest. I will finish the new typechecker and with it complete the Spiral language.

There is no obstacle that I cannot cross. The light of my inspiration will lead me to the end that I seek.