"4:45pm.     let apply_type d env_ty uniont b ret =         v_find env_ty b (fun () -> d.on_type_er d.trace <| sprintf "Cannot find a type named %s inside the type." b)             (fun ty -> inl " " (Op(TypeConstruct,[t uniont; t ty; v " "],None)))
No, I can't do it. It is simply not working. I feel as if I am in a swamp. The above can't possibly be right.

Also that thing with NamedT...the problems really start there.

But if I have NameT, do I really need named tuples in the first place?

At first I had the created at .Name(args). Then recently I got the bright idea of separating the Name from args, and being able to apply the Name with args to get the named tuple.

This would allow me to write brilliant code such as…
```
inl x = .Add (1,2)
arith x
```
In `inl x` the .Add of NameT and the plain tuple (1,2) would get combined to make .Add(1,2) which is a named tuple. The application would need to be overloaded for this.

But as I think how I would make the pattern matcher deal with this in nested expressions and now in type constructors I am completely freezing up.

For example, why not allow something like:
```
inl x = .Add 1 2
```
Well, why not? Because it is fucking complex that is why!

If I had to confess something, I felt that the original implementation of named tuples was already too complex as it was. What I have now is way over the line.

Instead of writing brilliant code as in the above, why not just do this instead.

```
inl f → f .Add 1 2
```

This would solve absolutely everything.

And instead of having named tuples, why not just write code such as:
```
typeinl
| .Add, x, y → x + y
| .Sub, x, y → x - y
```

5:25pm. Done with lunch.

Doing it like the above is brilliant. It is actually more concise without named tuples

And I can then expand that to partial application.

```
typeinl
| .Add x y → x + y
| .Sub x y → x - y
```"

The above is an excerpt from my journal. As strange as it is, despite working on this thing for like 4 months now, I still haven't gotten the named tuples down correctly.

There is also the issue with TyType. I'll talk about it tomorrow. Right now I am done for the day. I had to turn off for 1.5 hour since it had gotten stormy.