As usual, I made some progress, but now find myself completely perplexed again.
Adding the TypeContructorT type seemed like a good idea at first as it solves the immediate problem of how to treat the contructor types as void, but now that I look at it more closely, I find that I have no idea how to use it and in fact it makes types 10x more complicated.

...I am perplexed. Should I disallow returning of types after all? It seems like a reasonable course of action.

Agh, this is going to cause me some torment no doubt about it.

Let me take a break from coding so I can think about it. The plan is to do recursive types last after the F# backend is done.