Just this morning, I've realized I have a serious problem on my hands. That singleton tuple thing has finally become an issue for me.
Typeinl
| Int32 x → …
| Int64 x → …
| Float32 x → …

The problem is that for union types the above would not work. It would expect singleton tuples. So to match them I would have to write something like x :: (). In particular for pattern matching code, I think I should make parenthesis more sensitive and allow the (x) to match singletons. But still isn't there a convenient way to deal with this?

Ugh, these are the pains I have to deal with as a language designer. You would think that things like singleton tuples vs variables would be an easy thing to resolve, but it is not so. I have absolutely no idea what to do here.

But having to grip with things like these, clearly demonstrates I am back in action. It is nonsense like this that usually occupies my time.