I've added a tenth rule which states:
10) Singleton tuples returned from a type constructor case whose item's type is equal to the tuple's outer type, meaning the tuple is not a recursive or a union type will get implicitly converted to a variable. This so type construction application meshes with type constructor creation.

No matter what, it does not seem like I will be able to resolve this.
```
typeinl
| Int32 x → …
| Int64 x → …
| Float32 x → …
```
To bring up the previous example, passing a singleton tuple into the above will simply have to be an error even if the static data for it matches. I'll have to make sure to note this in the tutorial.

In addition to the above, I've combed the functions over and over in order to internalize them completely and now I think the semantics of it makes sense.

...Maybe.

`apply_typec` is something that I am going to have to test in order to really make sure.

Let me make the type constructor union and create functions. I need to refocus.