Ok, to start the day off I added the missing recursive type case into the `apply_typec`.
Taken together the `apply_typec` and `typec_create` functions are quite a bit bigger than the previous pattern matcher.

As a program, the Spiral compiler is quite a complex piece, definitely the hardest that I've ever made. The ML library will be kid's play after I am done with it.

Now...

I've plugged the two cases in. Is there anything else left to do before I move to work on the parser?