Ok, at this point I am basically too hungry to do anymore coding and I've been up since 7am at this. I think I will call it a day here.
That having said, I did get some insight just now. I do not need abstract interpretation here. I've been too focused on the forward way of thinking.

Instead I should start from the end and fold everything in. I do not think I even need the sequencer. Enough of that for now. I need to think about it for a bit.

But it is good that I started this.

To be honest, like usual, it feels like I did shit, but I did essentially make the first step towards completing this and that is important. Strangely enough, somehow 11 pages did appear in my journal as well which is about the usual for me.

I've started the file two months ago and already I am at 558 pages. It takes more than 5 seconds to save at this point.