I am having significant difficulty figuring out what to do here and am just thinking and iterating through the design ideas. As difficult as it is, I think I can settle on a design with some more time to think.
Right now what I am leaning towards is transforming the patterns into CPS form so that the partial evaluator can go through it in a single forward pass.

I actually know how to do the CPS transform and the bindings and the checks, but I need some more time to think. Some things still have not settled in for me yet.

Like do I need two arguments for the compiler function, one for success and one for failure or could I fit that somehow into the same closure?

Looking at it reasonably, yeah probably, but I still can't prove it to myself. I can't visualize the flow.

I have had a lot of difficulty focusing for the past week, but today despite not coding anything I am fully in the zone and homed in on the problem.

I need to resolve the desires swirling in me. Right now they are all mixing, but once they cool, they will be sharp as a blade.