"3:20pm. let rec compile_pattern arg pat (on_succ: Lazy<_>) (on_fail: Lazy<_>) =     match pat with     | PatVar (pos, x) -> l x arg pos on_succ.Value     | PatTuple (pos, l) ->         let len = List.length l         let on_succ =             List.foldBack (fun pat (s, i) ->                 let arg = tuple_index arg i pos                 lazy compile_pattern arg pat on_succ on_fail, i-1                 ) l (on_succ, len - 1)             |> fun (x,_) -> x
        case_ arg (lit_int len pos) on_succ.Value on_fail.Value pos

I just realized that this function does not need to be in the pevaler at all.

Oh, joy.

3:25pm. Right now I have a problem though. Just what the hell is that second argument to case doing?

It is completely useless.

And rather than destructuring those things….

3:50pm. Right now I am thinking about how to handle case.

After thinking on it for a while, I realized that actually the way I wrote the above is pretty much exactly correct. There isn't a better way of doing it.

What case should do is simple.

1) During peval, if the variable is a union type, it will destructure it a level into a tuple with a standard (non-union) type and branch on all of its cases.

2) If it is a primitive variable it will take the false branch of the case.

3) If it is tuple it will compare sizes. This rule is independent, but might be lead into from rule #1. Corollary to rule #1, tuples that have already been destructured will not be so again.

4) If the tuple sizes match, it will always take the true branch of the case.

5) If the sizes do not match, it will always take the false branch of the case.

Given how case will work, it might be necessary to hoist the clauses other than the first. Or better, hoist them all, but then inline those that are only used once.

No that won't work, I won't get the benefit of information passing that way since I can only do that at the end.

Crap."

The above is an excerpt from my journal.

Right now I am actually very close to nailing it. I just need to figure out how not to cause code bloat on repeat expressions brought in by case.

Well, in the worse case I suppose I could aways do it manually, but definitely I want something better.

...I think that for the rest of the day I want to just think.

The pattern compiler has like 12 lines right now, but as far as I am concerned it is mostly finished.

Once the design is decided upon the rest is just a matter of filling them in.